requirements.txt
astroid
attrs
autopep8
Click
Flask
gunicorn
importlib-metadata
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
more-itertools
packaging
pluggy
py
pycodestyle
PyJWT
pylint
pyparsing
pytest
python-dotenv
six
typed-ast
wcwidth
Werkzeug
wrapt
zipp
